更改循环索引类型并添加类型转换
将多个文件中的 `for` 循环索引变量从 `int` 类型更改为 `uint` 类型，以确保索引变量始终为非负数。
在 `SpvReflector.cs` 和 `VKGraphicsContext.PhysicalDevice.cs` 文件中，添加了类型转换 `(int)` 以确保索引变量在数组访问时正确转换为整数类型。

+++ src/ZenithEngine.ShaderCompiler/SpvReflector.cs
@@ -17,11 +17,11 @@ public static unsafe class SpvReflector
 
         List<ReflectResource> resources = [];
 
-        for (int i = 0; i < module->DescriptorSetCount; i++)
+        for (uint i = 0; i < module->DescriptorSetCount; i++)
         {
-            ReflectDescriptorSet set = module->DescriptorSets[i];
+            ReflectDescriptorSet set = module->DescriptorSets[(int)i];
 
-            for (int j = 0; j < set.BindingCount; j++)
+            for (uint j = 0; j < set.BindingCount; j++)
             {
                 DescriptorBinding* binding = set.Bindings[j];
